Moved test_create_table into separate file
